Versión 1.0.4, #452: Integrar ticket de configuración de resolución de menús y cambios necesarios en la base de datos.
git-svn-id: https://opengnsys.es/svn/branches/version1.0@2915 a21b9725-9963-47de-94b9-378ad31fedc9

+++ admin/Database/ogAdmBD.sql
@@ -436,7 +436,7 @@ CREATE TABLE IF NOT EXISTS `menus` (
   `grupoid` int(11) NOT NULL DEFAULT '0',
   `htmlmenupub` varchar(250) DEFAULT NULL,
   `htmlmenupri` varchar(250) DEFAULT NULL,
-  `resolucion` tinyint(4) DEFAULT NULL,
+  `resolucion` SMALLINT(4) DEFAULT NULL,
   PRIMARY KEY (`idmenu`)
 ) ENGINE=MyISAM  DEFAULT CHARSET=utf8 AUTO_INCREMENT=1 ;
